Completing module 25.

+++ 27_AdvancedRestAPI/server/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;  // checks whether its undefined
@@ -15,14 +16,14 @@ exports.getPosts = (req, res, next) => {
         .then(count => {
             totalItems = count;
             return Post.find()
-                       .skip((currentPage-1)*perPage)
-                       .limit(perPage);
+                .skip((currentPage - 1) * perPage)
+                .limit(perPage);
         })
         .then(posts => {
             res.status(200).json({
                 message: 'Fetched posts successfully.',
                 posts: posts,
-                totalItems:totalItems
+                totalItems: totalItems
             });
         })
         .catch(err => {
@@ -52,19 +53,31 @@ exports.postPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: 'Mohit'
+        creator: req.userId // we set this up in is-auth
     });                // In this we don't need to setup createAt since we added timestamps in the options.
 
     post.save()
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result
+                post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             });
         })
         .catch(err => {
@@ -85,7 +98,7 @@ exports.getPost = (req, res, next) => {
         .then(post => {
             if (!post) {
                 const error = new Error('Could not find post');
-                error.statusCode(404);
+                error.statusCode = 404;
                 throw error;   // if we throw an error it will we caught by the catch block.
             }
             return res.status(200).json({
@@ -121,7 +134,7 @@ exports.updatePost = (req, res, next) => {
 
     if (!imageUrl) {
         const error = new Error('No file picked');
-        error.statusCode(422);
+        error.statusCode = 422;
         throw error;
     }
 
@@ -129,7 +142,13 @@ exports.updatePost = (req, res, next) => {
         .then(post => {
             if (!post) {
                 const error = new Error('Could not find post');
-                error.statusCode(404);
+                error.statusCode = 404;
+                throw error;
+            }
+
+            if (post.creator.toString() !== req.userId.toString()) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
                 throw error;
             }
 
@@ -163,17 +182,30 @@ exports.deletePost = (req, res, next) => {
 
             if (!post) {
                 const error = new Error('Could not find post');
-                error.statusCode(404);
+                error.statusCode = 404;
                 throw error;
             }
 
             // Check logged in user.
+            if (post.creator.toString() !== req.userId.toString()) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
 
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            // Moongoose provides a pull method to remove an object from an array.            
+
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             res.status(200).json({ message: 'Deleted post' });
         })
         .catch(err => {
@@ -183,6 +215,64 @@ exports.deletePost = (req, res, next) => {
 
             next(err);
         })
+}
+
+exports.getStatus = (req, res, next) => {
+
+    User.findById(req.userId)
+        .then(user => {
+            if(!user)
+            {
+                const error=new Error('No user found');
+                error.statusCode=404;
+                throw error;
+            }
+
+            res.status(200).json({ status: user.status });
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+
+            next(err);
+        });
+}
+
+exports.updateStatus = (req, res, next) => {
+
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+        const error = new Error('Validation failed, entered data is incorrect.');
+        error.statusCode = 422;
+        throw error;  // will leave this function and refer to the next error handing part.
+    }
+    
+    const status = req.body.status;
+
+    User.findById(req.userId)
+        .then(user=>{
+            if(!user)
+            {
+                const error=new Error('No user found');
+                error.statusCode=404;
+                throw error;
+            }
+
+            user.status=status;
+            return user.save();
+        })
+        .then(user => {
+            res.status(200).json({ message:'Status updated' });
+        })
+        .catch(err => {
+            if (!err.statusCode) {
+                err.statusCode = 500;
+            }
+
+            next(err);
+        });
 }
